+++ app/src/main/java/com/mtkreader/MtkApplication.kt
@@ -4,8 +4,10 @@ import android.app.Application
 import android.content.Context
 import com.github.ivbaranov.rxbluetooth.RxBluetooth
 import com.mtkreader.contracts.DisplayDataContract
+import com.mtkreader.contracts.TimeContract
 import com.mtkreader.services.DisplayServiceImpl
 import com.mtkreader.services.ProcessDataServiceImpl
+import com.mtkreader.services.TimeServiceImpl
 import org.koin.core.context.startKoin
 import org.koin.dsl.module
 
@@ -16,6 +18,7 @@ class MtkApplication : Application() {
         // services
         single<DisplayDataContract.DisplayService> { DisplayServiceImpl() }
         single<DisplayDataContract.ProcessService> { ProcessDataServiceImpl() }
+        single<TimeContract.Service> { TimeServiceImpl() }
 
         // utils
         single { RxBluetooth(this@MtkApplication) }
