+++ app/src/main/res/values/strings.xml
@@ -239,4 +239,5 @@
 
     <string name="time_changed">Time has been changed</string>
     <string name="setting_time_is_impossible">Setting time is impossible! Set receiver in program mode!</string>
+    <string name="read_time">Read time</string>
 </resources>
